+++ packages/webapp/src/config/AppConfig.ts
@@ -25,7 +25,10 @@ export type MediaSources = {
   indexName?: string;
   [k: string]: unknown;
 }[];
-export type MediaViewDisableFlags = (("detail" | "map" | "similar" | "annotation" | "edit" | "tag") | "nav")[];
+export type MediaViewDisableFlags = (
+  | ("detail" | "map" | "similar" | "annotation" | "edit" | "tag" | "share")
+  | "nav"
+)[];
 
 /**
  * gallery configuration for webapp module